+++ src/Support/Contracts/RenderableOnMenu.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace CbtechLtd\Fastlane\Support\Contracts;
+
+use Illuminate\Support\Collection;
+
+interface RenderableOnMenu
+{
+    public function isVisibleOnMenu(): bool;
+
+    public function renderOnMenu(Collection $menu): void;
+}